pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_file = "src/fmu/sumo/uploader/_version.py"

[tool.isort]
profile = "black"

[tool.black]
line-length = 79

[project]
name = "fmu-sumo-uploader"
requires-python = ">=3.8"
dynamic = ["version"]
dependencies = [
    "sumo-wrapper-python>=1.0.3",
    "azure.storage.blob",
    "fmu-dataio",
    "httpx>=0.24.1",
    "OpenVDS; sys_platform != 'darwin' and python_version < '3.12'",
    "ert; sys_platform != 'win32'",
]

[project.optional-dependencies]
dev = ["black", "flake8", "pytest"]
test = ["pytest", "pytest-timeout", "fmu-sumo"]

docs = [
  "sphinx==6.2.1",
  "sphinx-rtd-theme",
  "autoapi",
  "sphinx-autodoc-typehints",
  "sphinxcontrib-apidoc",
]

[project.entry-points."ert"]
sumo_upload_forward_models = "fmu.sumo.uploader.hook_implementations.forward_models"

[project.entry-points."console_scripts"]
sumo_upload = "fmu.sumo.upload.forward_models.scripts.sumo_upload:main_entry_point"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true
platforms = ["any"]

[tool.setuptools.packages.find]
where = ["src"]

